Add serde for Configuration Ease of use

+++ src/structs.rs
@@ -1,19 +1,25 @@
+use std::path::PathBuf;
+use serde::{Serialize, Deserialize};
+
+#[derive(Serialize, Deserialize)]
 pub struct Configuration {
     pub name: String,
     pub callsign: String,
     pub maidenhead: String,
-    pub latitude: f64,
-    pub longitude: f64,
-    pub altitude: f64,
+    pub latitude: u64,
+    pub longitude: u64,
+    pub altitude: u64,
     pub logbooks: Vec<Logbook>,
 }
 
+#[derive(Serialize, Deserialize)]
 pub struct Logbook {
     pub callsign: String,
     pub log_name: String,
     pub entries: Vec<LogbookEntry>,
 }
 
+#[derive(Serialize, Deserialize)]
 pub struct LogbookEntry {
     pub datetime: String,
     pub station_callsign: String,
